+++ src/components/element_settings_builder.tsx
@@ -1,4 +1,4 @@
-import { ComponentSettings, IntegerSetting } from "./component_settings";
+import { ElementSettings, IntegerSetting } from "./element_settings";
 import { type LandingElement, type TextElement } from "./types"
 
 
@@ -6,14 +6,14 @@ type UpdateCallback = (updated: LandingElement) => void;
 
 
 const buildSettingsForTextElement = (component: TextElement, onUpdate: UpdateCallback) => {
-    return <ComponentSettings>
+    return <ElementSettings>
         <IntegerSetting
             name="Font size"
             value={component.styles?.fontSize || 10}
             min={10}
             onChange={value => onUpdate({ ...component, styles: { ...component.styles, fontSize: value }})}
         />
-    </ComponentSettings>
+    </ElementSettings>
 };
